Use greedy solving for branches
When the solution branches, we try solving the branch with fewest
errors. If that succeeds, we do not finish the remaining options.
There is one exception, if the remaining branches still have no
errors, we do continue. This to detect if there are multiple correct
solutions of the constraint problem, i.e. an ambiguity.