Add syntax proposal for relational definitions
To be able to define structural type relational definitions,
we propose this syntax for defining patterns on type relations.

The syntax consists of 2 additions:
  - A new RuleClause "forall" that can iterate over all occurrences
in a namespace. It defines the namespace, the variable name of
the occurrence and the defined type. It uses the declarationset
in a scope to retrieve all occurrences
  - A new relation definition which can take 2 arbitrary types
and a relation operator. The pattern matching construct also takes
a list of ruleclauses that will be generated when evaluating the
relation. This can take an arbitrary ruleclause, including the new
proposed "forall".

An example of this syntax is the following:

relations
    	sub: Type * Type {
    		UNION(asdasd) <sub! UNION(asds) := {
    			true.
    		},
    		RECORD(one) <sub! RECORD(other) := {
    			forall Field{d1}:ty1 in D(one) => {
    				Field{d1} -> other,
    				Field{d1} |-> d2,
    				d2 : ty2,
    				ty1 <sub? ty2.
				}.
			}
    	}

In this example the UNION type does not generate any constraints.
The RECORD type does which uses the forall ruleclause directly.