Remove special guard constraints from implementation.
As long as we have inequality in guards, delaying happens. Therefore,
solving guard constraints need to use the same fixed point mechanism as
other constraints. The implementation was actualy incorrect, the way it
was. If we drop inequality for pattern ordering to support fall-back
cases, we can simplify this again.