Add scope identity
Scope identity is created by using the trace of CUser constraints. Each
time a CUser constraint is resolved (not as a module boundary), the
cause applied to the new constraints will have the rule that was
executed. Whenever a scope is created with a CNew constraint, its cause
is used to generate a trace of rules. This trace is made of the
signatures of the rules invoked, i.e. the name of the rule as well as
the arguments of the rule to distinguish overloading. This is guaranteed
to be unique. In addition, the name of the variable is used in the name
of the scope, making it unique within a rule based on the specification
itself. Through this, scope identity is achieved. Each scope is unique
based on the name given in the specification, the trace describing
where it was created, and the module it resides in.