Add fake lazy debug context
It is a lazy debug context that is not actually lazy, used for the
modular solver to present log messages in order with system out and
system err. A faked lazy context was used to keep the mechanic of the
delayed log, which absorbs the messages, requiring messages to be
preserved in the first place.