Add single name query detection
From a query for a single name we can now determine the term that would
be matched by the query. This information is added to the query details
and can be used to create a index for names for dependencies.

For this, it was necessary to convert the patterns of the rule back into
iterms, which happens in SingleItemQuery. The different pattern classes
had to be made public in order to properly distinguish them.

The name class represents the occurrence of the queried item.