Split solvers
Split modules are an idea for limiting parent module changes by putting
the constraints of the parent that depend on other module in a separate
module, called the split module. The split module is then placed as a
child of the original module. This means that the original module cannot
be affected by changes to other modules.

The problems start however, when considering child modules. To create a
child module, the arguments must be ground. Whenever information for a
child module comes from other modules, it is created as a child of the
split module instead. I have not determined if such a scenario would be
wanted, or if it should be avoided, either by disallowing passing
non-local elements to the child module or otherwise.

Information from parent modules can be used, but queries on parent
scopes will become part of the split module. This is in line with the
want to create a stable module that does not change because of other
modules. The parent module itself cannot change, since it would be split
otherwise, but the query on the scope of the parent can be affected by
the children of the parent, including its split module.

The implementation as in this commit causes flaky solver behavior. This
is because delays are not on modules but on variables and edges. This
means that the solver order becomes important for if a constraint ends
up in the split module. The reason it sometimes fails is that
cross-module unification occurs from the split module. This can be
addressed as follows. Whenever a split module is created, all non ground
variables that are to be given a value should be substituted for a newly
created variable of the split module. In addition, the unifier of the
original module should be given an equality of the substituted variable
to the newly created variable.

The reason this is sound is because the original module will be done
solving and will not be able to execute any constraints that use the
given variable. It is not possible for the variable in question to cross
a module boundary, since variables have to be ground in order to be used
in module boundaries. This does give rise to the potential of creating
modules from split modules, which should generally be disallowed. Or
allowed and discouraged. These modules are susceptible to having to be
redone as their parent can change, and could complicate the model.

In this commit, the solving with split modules is not set up properly.
Currently, whenever a solver for an original module runs out of
constraints to solve, it creates a split module + solver with all the
constraints that are delayed at that point. However, this is not fully
correct. The solver should actively be prevented from using information
from other modules, and should instead delay on modules rather than
variables/edges.