Fix some of the split module issues
Split modules now no longer do cross module unification. Instead,
variables are created in the split module and new equality constraints
are added to the original module. These equality constraints are the
unification of the split variables to the non-split variables.

With this, split modules should work more as planned.

Keep in mind that as of this commit, the solver behavior has not been
adjusted to properly do split module solving.