Turn unification of rigid variables into delayed constraints
Instead of throwing an exception when a rigid variable is encountered,
collect these and return as part of the result. The user should decide
what they want to do with those. This is useful for applications that
want to do something with the remaining constraints, but requires the
rest to be normalized completely.