Implement disunifiction.
Add disunification to unification, discharging disequalities directly,
but keeping them around in the unifier in case they are violated later.

As a result, we also change the handling of rigid variables in
entailment. Instead of keeping unifications with rigid variables as
unsolved constraints -- which means they are not properly used un
unify/disunify reasoning -- they are now solved as normal. Entailments
checks if any were instantiated after the entailment solving is done.
Downside is that this is less efficient than the old technique.