Remove unifier from CExists again.
Having a unifier is not a proper replacement for substitution into
equals constraints. A simple example is applying `a :- f(a)`. Our
implementation would "apply" the substitution by adding an equality to
the unifier. This results in a recurisve unifier for this example. Also,
the eliminated variables fromt he domain of the substitution will remain
in the unifier, complicating many things related to capture avoidance
etc.